Add custom_vjp to simple functorch
Problem: users want to define functions with custom forward and backward
passes. These functions call PyTorch operations. When we vmap over such a
function, we would like for the backward pass to be preserved.

Why is this difficult? In PyTorch today, vmap over an autograd.Function
effectively runs vmap on the forward pass of the autograd.Function.
Meanwhile, autograd records the transformed operations for backward, instead
of the custom backward pass we specified in the autograd.Function!

Solution: this PR introduces a `custom_vjp` primitive that accepts
functions and varargs Tensor arguments and demonstrate that it resolves
the problem. My hope is that this PR gives us some insight into how we
could resolve the problem for functorch.